@@ -5,12 +5,12 @@ author: songyiYu<songyi0207@gmail.com>, yenoss<cpg0504@gmail.com>, FlutterMoum G
 homepage: https://github.com/flutter-moum/flutter_headset_event
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  mockito: 3.0.0
+  mockito: 5.0.10
 
 dev_dependencies:
   flutter_test:
